# notion-clone-api/openapi.et.yaml
openapi: 3.1.0
info:
  title: Ülesannete Haldamise API
  version: 1.0.0
  description: API ülesannete haldamiseks koos täieliku dokumentatsiooniga
  contact:
    name: Arendusmeeskond
    email: contact@yourdomain.com
servers:
  - url: /
    description: API serveri URL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT autentimise token
  schemas:
    Kasutajad:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
    PostUsersRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "email@domain.com"
        password:
          type: string
          format: password
          minLength: 6
          example: "minuparool"
      required:
        - email
        - password
    PatchUsersUserIdRequest:
      type: object
      properties:
        password:
          type: string
          format: password
          minLength: 6
          example: "uustugevõparool"
      required:
        - password
    Error:
      type: object
      properties:
        code:
          type: integer
          description: Vea HTTP staatuskood
          example: 400
        error:
          type: string
          description: Vea tüüp või kategooria
          example: Vigane Päring
        message:
          type: string
          description: Lühike kirjeldus, mis läks valesti
          example: Vigased sisendandmed
      required:
        - code
        - error
        - message
...
  responses:
    BadRequest:
      description: Vigane päring
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            BadRequest:
              value:
                code: 400
                error: "Vigane Päring"
                message: "Vigased sisendandmed"
    ForbiddenError:
      description: Juurdepääs keelatud
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            ForbiddenError:
              value:
                code: 403
                error: "Keelatud"
                message: "Teil pole sellele ressursile juurdepääsuõigust"
            InvalidTokenError:
              value:
                code: 403
                error: "Keelatud"
                message: "Vigane või aegunud token"
    InternalServerError:
      description: Serveri sisemine viga
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            InternalServerError:
              value:
                code: 500
                error: "Serveri Sisemine Viga"
                message: "Serveris tekkis ootamatu viga"
    NotFoundError:
      description: Ressurssi ei leitud
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            NotFoundError:
              value:
                code: 404
                error: "Ei Leitud"
                message: "Päritud ressurssi ei leitud"
    GoneError:
      description: Ressurss ei ole enam saadaval
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            GoneError:
              value:
                code: 410
                error: "Kadunud"
                message: "Päritud ressurss ei ole enam saadaval"
    UnauthorizedError:
      description: Autentimine on nõutav
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            UnauthorizedError:
              value:
                code: 401
                error: "Autoriseerimata"
                message: "Sellele ressursile juurdepääsuks on vaja autentimist"
...
paths:
  /users:
    post:
      tags:
        - Kasutajad
      summary: Loo uus kasutaja
      description: Registreeri uus kasutajakonto
      operationId: looKasutaja
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUsersRequest"
      responses:
        "201":
          description: Kasutaja edukalt loodud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kasutajad"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: E-post on juba olemas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                ConflictError:
                  value:
                    code: 409
                    error: "Konflikt"
                    message: "E-post on juba olemas"
    get:
      tags:
        - Kasutajad
      summary: Kuva kõik kasutajad
      description: Hangi nimekiri kõikidest kasutajatest
      operationId: hankiKasutajad
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: Kasutajate nimekiri
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kasutajad"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "410":
          $ref: "#/components/responses/GoneError"
  /users/{userId}:
    parameters:
      - name: kasutajaId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Kasutaja ID
    get:
      tags:
        - Kasutajad
      summary: Hangi kasutaja ID järgi
      description: Hangi üksikasjalik teave konkreetse kasutaja kohta
      operationId: hankiKasutajaIdJargi
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: Kasutaja andmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kasutajad"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "410":
          $ref: "#/components/responses/GoneError"
    delete:
      tags:
        - Kasutajad
      summary: Kustuta kasutajakonto
      description: Kustuta kasutaja konto ja kõik seotud ülesanded
      operationId: kustutaKasutaja
      security:
        - bearerAuth: [ ]
      responses:
        "204":
          description: Kasutaja edukalt kustutatud
          # 204 vastused ei tohiks sisaldada sisu
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "410":
          $ref: "#/components/responses/GoneError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags:
        - Kasutajad
      summary: Uuenda kasutajat
      description: Uuenda kasutaja teavet (praegu ainult parool)
      operationId: uuendaKasutajat
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchUsersUserIdRequest"
      responses:
        "200":
          description: Kasutaja edukalt uuendatud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kasutajad"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "410":
          $ref: "#/components/responses/GoneError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...